Events 100%

+++ ProftaakProject/Controllers/EvenementController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using ProftaakProject.Models;
 using ProftaakProject.Models.ConvertModels;
+using ProftaakProject.Models.Repositories;
 using ProftaakProject.Models.ViewModels.EventModels;
 using ProftaakProject.Repositories;
 
@@ -17,10 +18,12 @@ namespace ProftaakProject.Controllers
 
 
         private readonly EvenementRepo _evenementRepo;
+        private readonly UitzendRepo _uitzendRepo;
         private readonly EvenementToEvenementVMConverter _eevmc = new EvenementToEvenementVMConverter();
-        public EvenementController(EvenementRepo evr)
+        public EvenementController(EvenementRepo evr, UitzendRepo urepo)
         {
             _evenementRepo = evr;
+            _uitzendRepo = urepo;
         }
 
         public IActionResult Index()
@@ -50,7 +53,7 @@ namespace ProftaakProject.Controllers
         {
             Evenement ev = _evenementRepo.Read(id);
             EvenementViewModel evm = _eevmc.ConvertToViewModel(ev);
-            return View();
+            return View(evm);
         }
 
         [HttpGet]
@@ -67,7 +70,14 @@ namespace ProftaakProject.Controllers
             if (ModelState.IsValid)
             {
                 Evenement ev = _eevmc.ConvertToModel(evm);
-
+                if (evm.EnkelUitzendbureau)
+                {
+                    ev.Uitzendbureau = _uitzendRepo.GetByAccountID(GetUserId());
+                }
+                else
+                {
+                    ev.Uitzendbureau = new Uitzendbureau(-1);
+                }
                 _evenementRepo.Create(ev, GetUserId());
                 return RedirectToAction("Index");
             }
@@ -83,6 +93,7 @@ namespace ProftaakProject.Controllers
         }
 
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public IActionResult Edit(EvenementViewModel evm)
         {
             IActionResult retVal = View(evm);
@@ -108,7 +119,33 @@ namespace ProftaakProject.Controllers
         [HttpGet]
         public IActionResult Delete(int id)
         {
-            _evenementRepo.Delete(id);
+            if (_evenementRepo.Read(id).Host.Id == GetUserId())
+            {
+                _evenementRepo.Delete(id);
+            }
+
+            return RedirectToAction("Index");
+        }
+
+        [HttpGet]
+        public IActionResult SignOut(int id)
+        {
+            if (User.Identity.IsAuthenticated)
+            {
+                _evenementRepo.SignOut(id, GetUserId());
+            }
+            
+
+            return RedirectToAction("Index");
+        }
+
+        [HttpGet]
+        public IActionResult SignIn(int id)
+        {
+            if (User.Identity.IsAuthenticated)
+            {
+                _evenementRepo.SignIn(id, GetUserId());
+            }
 
             return RedirectToAction("Index");
         }
